API/requirements.txt
alembic==1.15.2
autopep8==2.3.2
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
google-ai-generativelanguage==0.4.0
google-api-core==2.24.2
google-auth==2.38.0
google-generativeai==0.3.2
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.62.3
gunicorn==21.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==24.2
proto-plus==1.26.1
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
SQLAlchemy==2.0.40
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.0.1
